Print help when executed directly with no args
If shaste is executed directly with no arguments (in otherwords, not
used with a pipe), the user likely doesn't know how to use it. Therefore
a help message for this case is a more useful default option than
reading text interactively and pasting it. For this more niche usecase,
consider `cat | shaste`